Eliminating more compiler warning on MSVC

+++ src/sagittarius/private/sagittariusdefs.h
@@ -329,7 +329,7 @@ typedef enum  {
 #define SG_INT_VALUE(obj)  (((long)SG_WORD(obj)) >> 2)
 #define SG_MAKE_INT(obj)   SG_OBJ((intptr_t)((long)((unsigned long)(obj) << 2) + 1))
 /* Do not use this!!! */
-#define SG_ENSURE_INT(obj) SG_OBJ((long)(obj) | 1)
+#define SG_ENSURE_INT(obj) SG_OBJ((intptr_t)((long)(intptr_t)(obj) | 1))
 #define SG_UINTP(obj)      (SG_INTP(obj)&&((long)SG_WORD(obj) >= 0))
 #define SG_INT_SIZE        (SIZEOF_LONG * 8 - 3)
 #define SG_INT_MAX         ((1L << SG_INT_SIZE) - 1)
